requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
crispy-bootstrap4==2024.10
cryptography==43.0.3
cssselect2==0.7.0
Django==4.2.18
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-js-asset==2.2.0
django-widget-tweaks==1.5.0
html5lib==1.1
idna==3.10
importlib_metadata==8.6.1
lxml==5.3.0
Markdown==3.7
oscrypto==1.3.0
pillow==11.1.0
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.2.0
PyPDF2==3.0.1
python-bidi==0.6.3
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
xhtml2pdf==0.2.16
zipp==3.21.0
